tests/requirements.txt
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.27.0
PyJWT==2.8.0
fastapi==0.109.0
